pyproject.toml
[project]
name = "dogdoctor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-multipart==0.0.6",
    "torch==2.1.1",
    "torchvision==0.16.1",
    "pillow==10.1.0",
    "pandas==2.1.3",
    "numpy==1.26.2",
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "python-dotenv==1.0.0",
    "sqlalchemy==2.0.23",
    "pymysql==1.1.0",
    "cv2",
]
[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

